third_party/blink/web_tests/http/tests/websocket/broken-utf8-expected.txt

Test whether WebSocket aborts the connection when it receives a text frame containing broken UTF-8 data.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
onerror() was called.
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
